requirements.txt
android-backup==0.2.0
appdirs==1.4.3
asn1crypto==0.24.0
attrs==18.2.0
cffi==1.12.2
Click==7.0
construct==2.9.45
cryptography==2.4.1
idna==2.7
ifaddr==0.1.4
ipaddress==1.0.22
pretty-cron==1.2.0
prometheus-client==0.4.2
pycparser==2.19
python-miio==0.5.0.1
pytz==2018.7
six==1.11.0
tqdm==4.28.1
zeroconf==0.21.3
python-aqi==0.6.1
